requirements.txt
tensorflow
streamlit
pillow
opencv-python
scikit-learn
matplotlib
seaborn
